Refactored to decouple recovery model build code.
